move the video tests in a single file
