+++ tests-e2e/ocw-ci-test-course/video-tabs.spec.ts
@@ -1,6 +1,114 @@
 import { test, expect } from "@playwright/test"
-import { VideoElement } from "../util/VideoElement"
 import { CoursePage } from "../util"
+import { VideoElement } from "../util/VideoElement"
+
+test("Verify that the 'Download video' and 'Download transcript' links are keyboard navigable and have the correct download URLs", async ({
+  page
+}) => {
+  const coursePage = new CoursePage(page, "course")
+  await coursePage.goto("resources/ocw_test_course_mit8_01f16_l01v01_360p")
+  const downloadLinks = [
+    "https://live-qa.ocw.mit.edu/courses/123-ocw-ci-test-course-fall-2022/ocw_test_course_mit8_01f16_l01v01_360p_360p_16_9.mp4",
+    "https://live-qa.ocw.mit.edu/courses/8-01sc-classical-mechanics-fall-2016/33f61131009a6cd12d9a4c0e42eb7f44_ErlP_SBcA1s.pdf"
+  ]
+  const downloadButton = page.getByRole("button", {
+    name: `Download Video and Transcript`
+  })
+  await downloadButton.focus()
+  await page.keyboard.press("Enter")
+
+  const videoDownloadLink = page.getByRole("link", { name: "Download video" })
+  const transcriptDownloadLink = page.getByRole("link", {
+    name: "Download transcript"
+  })
+  const downloadLinksArr = [videoDownloadLink, transcriptDownloadLink]
+
+  for (let i = 0; i < 2; i++) {
+    await expect(downloadLinksArr[i]).toBeVisible()
+    await page.keyboard.press("Tab")
+    const hrefAttribute = await page.locator(":focus").getAttribute("href")
+    expect(hrefAttribute).toBe(downloadLinks[i])
+  }
+})
+
+test("Embed video redirects to video page using keyboard navigation", async ({
+  page
+}) => {
+  const coursePage = new CoursePage(page, "course")
+  await coursePage.goto("pages/video-series-overview/")
+  const videoRedirectLink = page.getByRole("link", {
+    name: "View video page"
+  })
+  await videoRedirectLink.focus()
+  page.keyboard.press("Enter")
+  await coursePage.page.waitForURL(
+    "**/resources/ocw_test_course_mit8_01f16_l01v01_360p/"
+  )
+  expect(coursePage.page.url()).toContain(
+    "resources/ocw_test_course_mit8_01f16_l01v01_360p"
+  )
+})
+test("Video tabs content (links) are keyoard navigable", async ({ page }) => {
+  const tabs = [
+    {
+      title: "Related Resources",
+      url:   "courses/ocw-ci-test-course/resources/example_pdf/"
+    },
+    {
+      title: "Optional Tab",
+      url:   "courses/ocw-ci-test-course/resources/example_notes/"
+    }
+  ]
+  for (const tab of tabs) {
+    const coursePage = new CoursePage(page, "course")
+    await coursePage.goto("resources/ocw_test_course_mit8_01f16_l01v01_360p")
+    const tabButton = page.getByRole("button", {
+      name: `${tab.title}`
+    })
+    await tabButton.focus()
+    page.keyboard.press("Enter")
+    page.keyboard.press("Tab")
+    page.keyboard.press("Enter")
+    await coursePage.page.waitForURL(`**/${tab.url}`)
+    expect(coursePage.page.url()).toContain(tab.url)
+  }
+})
+test.only("Expand and collapse video tabs using keyboard navigation", async ({
+  page
+}) => {
+  const coursePage = new CoursePage(page, "course")
+  await coursePage.goto("resources/ocw_test_course_mit8_01f16_l01v01_360p")
+  const videoPage = new VideoElement(page)
+
+  const tabClassToTitle: { [key: string]: string } = {
+    transcript:         "Transcript",
+    "related-resource": "Related Resources",
+    "optional-tab":     "Optional Tab"
+  }
+
+  for (const [tabClass, tabTitle] of Object.entries(tabClassToTitle)) {
+    const tabButton = page.getByRole("button", {
+      name:  `${tabTitle}`,
+      exact: true
+    })
+    await expect(tabButton).toBeVisible()
+    await tabButton.focus()
+
+    page.keyboard.press("Enter")
+    expect(
+      await videoPage.tab({ name: `${tabTitle}`, expanded: true }).waitFor()
+    )
+    // Wait for the tab to fully expand
+    await expect(
+      page.locator(`.video-tab.container.${tabClass}.collapse`)
+    ).toHaveClass(/show/)
+    // Collapse the tab
+    page.keyboard.press("Enter")
+    expect(
+      await videoPage.tab({ name: `${tabTitle}`, expanded: false }).waitFor()
+    )
+  }
+})
 
 test("A page without a transcript has the proper tab titles and contents", async ({
   page
@@ -10,7 +118,7 @@ test("A page without a transcript has the proper tab titles and contents", async
   const videoPage = new VideoElement(page)
 
   await expect(videoPage.tab({})).toHaveText([
-    /\s*$/, // first tab is just download button with no title.
+    /Transcript/,
     /Related Resources/,
     /Optional Tab/
   ])
@@ -30,7 +138,7 @@ test("A page without a transcript has the proper tab titles and contents", async
     '<a href="/courses/ocw-ci-test-course/resources/example_pdf/">(PDF)</a>'
   )
   expect(optionalTabHTML).toContain(
-    '<a href="https://ocw.mit.edu" target="_blank" rel="noopener">OCW</a>'
+    '<a href="/courses/ocw-ci-test-course/resources/example_notes/">(PDF)</a>'
   )
 })
 
@@ -38,5 +146,5 @@ test("A page with a transcript has a transcript tab", async ({ page }) => {
   const coursePage = new CoursePage(page, "course")
   await coursePage.goto("/resources/ocw_test_course_mit8_01f16_l01v02_360p")
   const videoPage = new VideoElement(page)
-  await expect(videoPage.tab({})).toHaveText(["Transcript"])
+  await expect(videoPage.tab({})).toHaveText(/Transcript\s*/)
 })
